Biome Base

+++ src/main/java/cn/nukkit/level/generator/noise/Simplex.java
@@ -47,15 +47,15 @@ public class Simplex extends Perlin {
     }
 
 
-    protected static double dot2D(double[] g, double x, double y) {
+    protected static double dot2D(int[] g, double x, double y) {
         return g[0] * x + g[1] * y;
     }
 
-    protected static double dot3D(double[] g, double x, double y, double z) {
+    protected static double dot3D(int[] g, double x, double y, double z) {
         return g[0] * x + g[1] * y + g[2] * z;
     }
 
-    protected static double dot4D(double[] g, double x, double y, double z, double w) {
+    protected static double dot4D(int[] g, double x, double y, double z, double w) {
         return g[0] * x + g[1] * y + g[2] * z + g[3] * w;
     }
